BaseAjaxAddView
