+++ bika/lims/browser/analysisrequest/add2.py
@@ -5,14 +5,10 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from datetime import datetime
-
 import magnitude
-from DateTime import DateTime
 from Products.CMFPlone.utils import _createObjectByType
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from plone import protect
 from plone.memoize.volatile import cache
 from zope.interface import implements
 from zope.publisher.interfaces import IPublishTraverse
@@ -21,9 +17,9 @@ from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
 from bika.lims.browser.base_add_view import BaseAddView
+from bika.lims.browser.base_add_view import BaseAjaxAddView
 from bika.lims.browser.base_add_view import BaseManageAddView
 from bika.lims.utils import cache_key
-from bika.lims.utils import returns_json
 from bika.lims.utils import tmpID
 from bika.lims.utils.analysisrequest import create_analysisrequest as crar
 
@@ -458,125 +454,14 @@ class AnalysisRequestManageView(BaseManageAddView):
         return self.tmp_obj
 
 
-class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
+class ajaxAnalysisRequestAddView(BaseAjaxAddView, AnalysisRequestAddView):
     """Ajax helpers for the analysis request add form
     """
     implements(IPublishTraverse)
 
     def __init__(self, context, request):
-        super(ajaxAnalysisRequestAddView, self).__init__(context, request)
-        self.context = context
-        self.request = request
-        self.traverse_subpath = []
-        # Errors are aggregated here, and returned together to the browser
-        self.errors = {}
-
-    def publishTraverse(self, request, name):
-        """ get's called before __call__ for each path name
-        """
-        self.traverse_subpath.append(name)
-        return self
-
-    @returns_json
-    def __call__(self):
-        """Dispatch the path to a method and return JSON.
-        """
-        protect.CheckAuthenticator(self.request.form)
-        protect.PostOnly(self.request.form)
-
-        if len(self.traverse_subpath) != 1:
-            return self.error("Not found", status=404)
-        func_name = "ajax_{}".format(self.traverse_subpath[0])
-        func = getattr(self, func_name, None)
-        if func is None:
-            return self.error("Invalid function", status=400)
-        return func()
-
-    def error(self, message, status=500, **kw):
-        """Set a JSON error object and a status to the response
-        """
-        self.request.response.setStatus(status)
-        result = {"success": False, "errors": message}
-        result.update(kw)
-        return result
-
-    def to_iso_date(self, dt):
-        """Return the ISO representation of a date object
-        """
-        if dt is None:
-            return ""
-        if isinstance(dt, DateTime):
-            return dt.ISO8601()
-        if isinstance(dt, datetime):
-            return dt.isoformat()
-        raise TypeError("{} is neiter an instance of DateTime nor datetime"
-                        .format(repr(dt)))
-
-    def get_records(self):
-        """Returns a list of AR records
-
-        All fields coming from `request.form` have a number prefix, e.g. `Contact-0`.
-        All fields with the same suffix number are grouped together in a record.
-        Each record represents the data for one column in the AR Add form and
-        contains a mapping of the fieldName (w/o prefix) -> value.
-
-        Example:
-        [{"Contact": "Rita Mohale", ...}, {Contact: "Neil Standard"} ...]
-        """
-        form = self.request.form
-        ar_count = self.get_obj_count()
-
-        records = []
-        # Group belonging AR fields together
-        for arnum in range(ar_count):
-            record = {}
-            s1 = "-{}".format(arnum)
-            keys = filter(lambda key: s1 in key, form.keys())
-            for key in keys:
-                new_key = key.replace(s1, "")
-                value = form.get(key)
-                record[new_key] = value
-            records.append(record)
-        return records
-
-    def get_uids_from_record(self, record, key):
-        """Returns a list of parsed UIDs from a single form field identified by the given key.
-
-        A form field ending with `_uid` can contain an empty value, a
-        single UID or multiple UIDs separated by a comma.
-
-        This method parses the UID value and returns a list of non-empty UIDs.
-        """
-        value = record.get(key, None)
-        if value is None:
-            return []
-        if isinstance(value, basestring):
-            value = value.split(",")
-        return filter(lambda uid: uid, value)
-
-    def get_objs_from_record(self, record, key):
-        """Returns a mapping of UID -> object
-        """
-        uids = self.get_uids_from_record(record, key)
-        objs = map(self.get_object_by_uid, uids)
-        return dict(zip(uids, objs))
-
-    @cache(cache_key)
-    def get_base_info(self, obj):
-        """Returns the base info of an object
-        """
-        if obj is None:
-            return {}
-
-        info = {
-            "id": obj.getId(),
-            "uid": obj.UID(),
-            "title": obj.Title(),
-            "description": obj.Description(),
-            "url": obj.absolute_url(),
-        }
-
-        return info
+        AnalysisRequestAddView.__init__(self, context, request)
+        BaseAjaxAddView.__init__(self, context, request)
 
     @cache(cache_key)
     def get_client_info(self, obj):
@@ -937,14 +822,6 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
         info.update({})
         return info
 
-    @cache(cache_key)
-    def get_preservation_info(self, obj):
-        """Returns the info for a Preservation
-        """
-        info = self.get_base_info(obj)
-        info.update({})
-        return info
-
     def get_service_partitions(self, service, sampletype):
         """Returns the Partition info for a Service and SampleType
 
@@ -990,15 +867,6 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             })
 
         return partitions
-
-    def ajax_get_global_settings(self):
-        """Returns the global Bika settings
-        """
-        bika_setup = api.get_bika_setup()
-        settings = {
-            "show_prices": bika_setup.getShowPrices(),
-        }
-        return settings
 
     def ajax_get_service(self):
         """Returns the services information
